requirements.txt
xgboost
torch
numpy
pandas
scikit-learn
pytorch-lightning
matplotlib
seaborn
mlxtend
openpyxl
albumentations
PyYAML